Refactor: asset 페이지 컴포넌트 구조 변경 및 import 경로 수정
