+++ src/pages/asset/AssetMain.vue
@@ -12,7 +12,7 @@
 <script setup>
 import { ref } from 'vue';
 
-import AssetTabSwitcher from './common/AssetTabSwitcher.vue'; //상단 탭
+import AssetTabSwitcher from './component/common/AssetTabSwitcher.vue'; //상단 탭
 
 //각 탭별 화면
 import AssetTotalTab from './tabs/AssetTotalTab.vue';
