+++ src/pages/asset/component/account/AccountDetailPage.vue
