+++ src/pages/asset/component/common/AddItemButton.vue
