+++ src/pages/asset/component/common/AssetTabSwitcher.vue
