+++ src/pages/asset/component/common/NoDataCard.vue
