+++ src/pages/asset/component/total/TotalAssetCard.vue
